+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Minor Improvements
 
 * Tabs emits a onTabClick event when on the headers is clicked
+* Add phone, email, location and mobile icons
 
 ## Bug Fixes